New Office Room + New Camera based on previous script
I made a new office room to hopefully make adding textures and everything a bit better when we advance more in the project.

A new camera script was also created using the old script, it just makes it so boundaries can be set in the X and Z direction, which will not allow the user to move the camera view out of the room (if the correct values are chosen of course)

There are also more textures that were added, along with my attempt at creating a safe object for us to use.