making code more resuable
there is a MVC architecture design with the folders for our scripts.

The camera_view script can be applied to any of the room's camera's as long as the camera_controller gets updated with the things it needs to accomdate a new room.

This only applies to the rooms so far, unsure about the minigames